Fix bug in probe_28sf040() causing flash corruption on SST49LF160C verify
The first byte of the flash chip was read at the start of the function
and later written back to address 0 if the flash chip was not identified
as SST28SF040, which means most of the time. This write caused corruption
of flash contents when verifying a SST49LF160C part.

Corresponding to flashrom svn r126 and coreboot v2 svn r2744.

Signed-off-by: Ed Swierk <eswierk@arastra.com>
Acked-by: Peter Stuge <peter@stuge.se>