I looked at the datasheet and erase_sector_39sf020() is totally and completely wrong
It was a straight cut'n'paste from SST 28SF040 code and the person doing
the cut'n'paste didn't even bother to check the data sheet. The SST 39SF020
is completely incompatible with the 28SF040. No need for replacement.
According to the data sheet, standard JEDEC commands will work and we have
those commands in the tree already.

Corresponding to flashrom svn r242 and coreboot v2 svn r3331.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Joseph Smith <joe@settoplinux.org>