Add additional SPI sector erase and chip erase command functions to flashrom
Not all chips support all commands, so allow the implementer to select
the matching function. Fix a layering violation in ICH SPI code to be
less bad. Still not perfect, but the new code is shorter, more generic and
architecturally more sound. TODO (in a separate patch): - move the generic
sector erase code to spi.c - decide which erase command to use based on info
about the chip - create a generic spi_erase_all_sectors function which calls
the generic sector erase function Thanks to Stefan for reviewing and commenting.

Corresponding to flashrom svn r337 and coreboot v2 svn r3722.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>