When flashrom JEDEC code sends the ID command to the chip, it expects to see IDs in the default flash location
However, sometimes the chip does not react to the ID command, either because
it doesn't understand the command or because the command never reached it.
One way to detect this is to compare ID output with flash chip contents for
the same location. If they are identical, there is a high chance you're not
actually seeing ID output. Warn the user in that case. This patch helps a lot
when a chip is not recognized and we want to check if the probe responses are
real IDs or just random flash chip contents. This should probably be added
to all probe functions, but probe_jedec is called for all sizes and thus
flashrom will check this condition at least once per size, making sure we
can cross-match the warning.

Corresponding to flashrom svn r494.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: FENG Yu Ning <fengyuning1984@gmail.com>