There are various reasons why a SPI command can fail
Among others, I have seen the following problems: - The SPI opcode is not
supported by the controller. ICH-style controllers exhibit this if SPI config
is locked down. - The address in in a prohibited area. This can happen on ICH
for any access (BBAR) and for writes in chipset write protected areas. - There
is no SPI controller. Introduce separate error codes for unsupported opcode
and prohibited address. Add the ability to adjust REMS and RES addresses to
the minium supported read address with the help of spi_get_valid_read_addr().
That function needs to call SPI controller specific functions like reading
BBAR on ICH.

Corresponding to flashrom svn r500.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>