Current programmer parameter syntax is -p programmer=parameter
Unfortunately, many parameters are of the form variable=val, so we get
commandlines like this: flashrom -p it87spi=port=0x820 and this looks horrible.
Using : instead of = would make such parameters look better: flashrom -p
it87spi:port=0x820 As a side benefit, this patch mentions the programmer name
in the error message if it is unknown.

Corresponding to flashrom svn r693.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Stefan Reinauer <stepan@coresystems.de>