The current ICH SPI preop handling is a hack which spews lots of warnings, but still yields correct results
With the multicommand infrastructure I introduced in r645, it became possible
to integrate ICH SPI preopcodes cleanly into the flashrom design. The new
code checks for every opcode in a multicommand array if it is a preopcode.
If yes, it checks if the next opcode is associated with that preopcode and
in that case it simply runs the opcode because the correct preopcode will be
run automatically before the opcode.

Corresponding to flashrom svn r727.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: FENG Yu Ning <fengyuning1984@gmail.com>