Since we don't have any debug level printing infrastructure yet, I propose to kill the obnoxious debug message in ichspi.c which was added
To check for correct PREOP handling. We know the code works fine (after
getting a few reports over 100 MB long) and there's no point in keeping it
around anymore. If there is any desire, we can reinstate it as print_spew
or whatever once the debug level infrastructure is merged, but at that point
we probably just are happy that the debug output isn't there anymore.

Corresponding to flashrom svn r753.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Luc Verhaegen <libv@skynet.be>