If a chip is not on the RDID generic vendor list nor on the REMS specific ID list, flashrom will claim that no chip is there
Handle these cases gracefully. flashrom will ignore generic matches if a
specific chip was found, so this will have no impact on supported chips,
but help a lot for a first quick analysis by the user or developer. The only
drawback is that unknown chips may be recognized multiple times until they
are added to flashchips.[ch].

Corresponding to flashrom svn r767.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Marc Jones <marcj303@gmail.com>