Add the ability to generate test patterns for write testing
This will be useful once we create a--test function for flashrom. The test
patterns make it easy to find skipped and duplicated bytes, are human readable,
and the first 8 of them have block numbers to detect aliasing or wraparounds.
Current size limit for aliasing detection is 16 MByte, but since neither
LPC nor FWH nor SPI chips exist with bigger sizes, this is reasonably safe.
Detailed documentation is available as source code comments above the new
function generate_testpattern().

Corresponding to flashrom svn r770.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Maciej Pijanka <maciej.pijanka@gmail.com>