Fix the DOS port
Now the DS selector limit is set to 4GB and all mmio accesses goes through DS,
the 1:1 mapping is fixed so the _DS base is taken onto account. Plus is that
the hwaccess.c needs no change and memcpy etc can be used on mmaped space.

Corresponding to flashrom svn r995.

Signed-off-by: Rudolf Marek <r.marek@assembler.cz> 
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>