Convert all PCI-based external programmers to use special little-endian accessors for all MMIO regions of PCI devices
This patch does _not_ touch the internal programmer (which is PCI-based
as well). Huge thanks go to Misha Manulis who worked with me to create a
first version of this patch for the satasii programmer based on modification
of generic code. Huge thanks also go to Segher Boessenkool for suggesting
the pci_mmio_ prefix for the abstraction layer. NOTE to package maintainers:
With this patch, compilation and usage of flashrom should be safe on x86,
x86_64, MIPS (little and big endian) and PowerPC (big endian). The internal
programmer is disabled on non-x86/x86_64 (but it compiles). The atahpt,
nic3com, nicnatsemi, nicrealtek and rayer_spi can not be compiled on
non-x86/x86_64 because port space I/O is not (yet) supported. Please compile
with default settings on x86/x86_64 and with the following settings on
all other architectures: make CONFIG_NIC3COM=no CONFIG_NICREALTEK=no
CONFIG_NICNATSEMI=no CONFIG_RAYER_SPI=no

Corresponding to flashrom svn r1111.

Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Misha Manulis <misha@manulis.com>