Implement libpayload support and improve life for DOS based flashrom, too
Change the physmap* behaviour to use (void*)-1 as error code instead of NULL.
That way, 1:1 mapped memory can be supported properly because (void*)0 is not
a magic pointer anymore. (void*)-1 on the other hand is a rather unlikely
memory offset, so that should be safe.

Corresponding to flashrom svn r1181.

Signed-off-by: Patrick Georgi <patrick.georgi@coresystems.de>
Signed-off-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>