print_wiki: Fixes, explanation re
Flash chips, links to footnotes. - Escape mainboard notes in the wiki output
with <nowiki> (this became necessary because MediaWiki started to convert
<mark@zl2tod.net> to <mark>). - Add explanations at the top of the flash chip
table in the wiki. - Add links to the footnotes of mainboards and laptops
similarly to what MediaWiki does: link from the reference to the footnote
and vice versa too. Also, check for truncations of the footnotes due to the
limited buffer. - Fix a typo introduced in r1555.

Corresponding to flashrom svn r1694.

Signed-off-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>
Acked-by: Stefan Tauner <stefan.tauner@alumni.tuwien.ac.at>