Add support for PPC LPC2SPI SPI host controller.
This controller sits on the CPU's LPC bus on POWER8 systems and acts as a SPI
controller with "mmio" and command-based register spaces.

Similar code exists in the (Apache license) skiboot open-power project here:
https://github.com/open-power/skiboot/blob/master/hw/sfc-ctrl.c

This implementation is my own work and was not derived from the skiboot code
(in fact the skiboot code was based on this).

Change-Id: I9a6014b03fba406f8f2cb993e27ae73bd2c6b0d1
Signed-off-by: Robert Lippert <rlippert@google.com>